bug & explore: compute normalization factor over absolute value of fluxes. Add notebook to explore anomalies from u band
In the past the normalization factor was computed over the raw time series. When we kkep negative fluxes, we can get a negative mean. Now we set the mean value over the absolute value of fluxes to avoid this situation. The notebook to explore anomalies in the u band handles the histogram and mean distance between ligh curves