explore: test dependence of distances with dinamilcal range between two time series
In the amplitude_distance notebook, we take two series, mask the negative fluxes and compute distances keeping the amplitude on one constant and changing the other playing around with the orders of magnitude. The difference in amplitude between the original series is of 2 orders of manitude. The difference is very substantial