feature: add diverse normalization fuctionality and masking of negative fluxes
The normalizations are: mean, max, min_max. The min_max normalization computes the distance between the min and max values and normalizes by it. The max normalization gets the max of the absolute values in fluxes. The mean normalization takes the mean on the absolute values of the flux. The masking discards negative fluxes :)