explore: inspect descriptive stats of agn lcs
Explore the distribution of the min, max, max-min, cadence_count and median magnitude error of the data set. Over all the data set is well binned for Wasserstein and Bottleneck distance. The only factor that might bias this scores is the cadence_count (number of measurements in each time series) since it spreads from 40 to 180 points with a mean of ~ 100